Added ability to build static library for iOS and Fabric support.

+++ Framework/scripts/build_universal.sh
@@ -2,24 +2,32 @@
 
 set -e
 
+FRAMEWORKS_BUILD_CONFIGURATION="Release"
 FRAMEWORKS_PATH="${BUILD_DIR}/Frameworks"
-[[ "${TARGET_NAME}" =~ (iOS|watchOS) ]] && FRAMEWORK_NAME="PubNub (${BASH_REMATCH[1]})"
+FRAMEWORKS_TARGET_ARCHITECTURES='i386 x86_64'
+[[ "${TARGET_NAME}" =~ (iOS|watchOS|Fabric) ]] && FRAMEWORK_NAME="PubNub (${BASH_REMATCH[1]})"
+[[ ${BUILDING_STATIC_FRAMEWORK:=0} == 1 ]] && FRAMEWORK_NAME="Static ${FRAMEWORK_NAME}"
+
 PRODUCTS_PATH="${SRCROOT}/Products"
 
 # Clean up from previous builds
-if [[ -d "${FRAMEWORKS_PATH}" ]]; then
-    rm -R "${FRAMEWORKS_PATH}"
-fi
-if [[ -d "${PRODUCTS_PATH}" ]]; then
-    rm -R "${PRODUCTS_PATH}"
-fi
+[[ -d "${FRAMEWORKS_PATH}" ]] && rm -R "${FRAMEWORKS_PATH}"
+[[ -d "${PRODUCTS_PATH}" ]] && rm -R "${PRODUCTS_PATH}"
 
 PLATFORMS=($SUPPORTED_PLATFORMS)
 # Compile framework for all required platforms.
 for sdk in "${PLATFORMS[@]}"
 do
     echo "Building ${FRAMEWORK_NAME} for ${sdk}..."
-    xcrun --no-cache xcodebuild -project "${PROJECT_FILE_PATH}" -target "${FRAMEWORK_NAME}" -configuration "${CONFIGURATION}" -sdk "${sdk}" BUILD_DIR="${BUILD_DIR}" OBJROOT="${OBJROOT}" BUILD_ROOT="${BUILD_ROOT}" SYMROOT="${SYMROOT}" ONLY_ACTIVE_ARCH=NO $ACTION > /dev/null
+    if [[ $sdk =~ (simulator) ]]; then
+        [[ "${sdk}" =~ (iphone) ]] && FRAMEWORKS_TARGET_ARCHITECTURES="i386 x86_64"
+        [[ "${sdk}" =~ (watch) ]] && FRAMEWORKS_TARGET_ARCHITECTURES="i386"
+    else
+        [[ "${sdk}" =~ (iphone) ]] && FRAMEWORKS_TARGET_ARCHITECTURES="armv7 armv7s arm64"
+        [[ "${sdk}" =~ (watch) ]] && FRAMEWORKS_TARGET_ARCHITECTURES="armv7k"
+    fi
+
+    xcrun --no-cache xcodebuild -project "${PROJECT_FILE_PATH}" -target "${FRAMEWORK_NAME}" -configuration "${FRAMEWORKS_BUILD_CONFIGURATION}" -sdk "${sdk}" BUILD_DIR="${BUILD_DIR}" OBJROOT="${OBJROOT}" BUILD_ROOT="${BUILD_ROOT}" SYMROOT="${SYMROOT}" ARCHS="${FRAMEWORKS_TARGET_ARCHITECTURES}" VALID_ARCHS="${FRAMEWORKS_TARGET_ARCHITECTURES}"  ONLY_ACTIVE_ARCH=NO $ACTION > /dev/null
     echo "Built ${FRAMEWORK_NAME} for ${sdk}"
 done
 
@@ -30,9 +38,9 @@ BUILT_FRAMEWORKS=("CocoaLumberjack" "PubNub")
 for sdk in "${PLATFORMS[@]}"
 do
     if [[ $sdk =~ (simulator) ]]; then
-        SIMULATOR_ARTIFACTS_PATH="${BUILD_DIR}/${CONFIGURATION}-${sdk}"
+        SIMULATOR_ARTIFACTS_PATH="${BUILD_DIR}/${FRAMEWORKS_BUILD_CONFIGURATION}-${sdk}"
     else
-        OS_ARTIFACTS_PATH="${BUILD_DIR}/${CONFIGURATION}-${sdk}"
+        OS_ARTIFACTS_PATH="${BUILD_DIR}/${FRAMEWORKS_BUILD_CONFIGURATION}-${sdk}"
     fi
 done
 
@@ -43,11 +51,35 @@ mkdir -p "${PRODUCTS_PATH}"
 # Copy ARM binaries and build "fat" binary for each built framework.
 for frameworkName in "${BUILT_FRAMEWORKS[@]}"
 do
-    FRAMEWORK_BUNDLE_NAME="${frameworkName}.framework"
-    FRAMEWORK_ARM_BUILD_PATH="${OS_ARTIFACTS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
-    FRAMEWORK_SIM_BUILD_PATH="${SIMULATOR_ARTIFACTS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
-    FRAMEWORK_DESTINATION_PATH="${FRAMEWORKS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
-    cp -r "${FRAMEWORK_ARM_BUILD_PATH}" "${FRAMEWORK_DESTINATION_PATH}"
-    xcrun lipo -create "${FRAMEWORK_DESTINATION_PATH}/${frameworkName}" "${FRAMEWORK_SIM_BUILD_PATH}/${frameworkName}" -output "${FRAMEWORK_DESTINATION_PATH}/${frameworkName}"
-    cp -r "${FRAMEWORKS_PATH}/${FRAMEWORK_BUNDLE_NAME}" "${PRODUCTS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
+    if [[ ${BUILDING_STATIC_FRAMEWORK:=0} == 0 ]]; then
+        FRAMEWORK_BUNDLE_NAME="${frameworkName}.framework"
+        FRAMEWORK_ARM_BUILD_PATH="${OS_ARTIFACTS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
+        FRAMEWORK_SIM_BUILD_PATH="${SIMULATOR_ARTIFACTS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
+        FRAMEWORK_DESTINATION_PATH="${FRAMEWORKS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
+    else
+        FRAMEWORK_BUNDLE_NAME="PubNub.framework"
+        FRAMEWORK_ARM_BUILD_PATH="${OS_ARTIFACTS_PATH}"
+        FRAMEWORK_SIM_BUILD_PATH="${SIMULATOR_ARTIFACTS_PATH}"
+        FRAMEWORK_DESTINATION_PATH="${FRAMEWORKS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
+    fi
+
+    FRAMEWORK_PRODUCTS_PATH="${PRODUCTS_PATH}/${FRAMEWORK_BUNDLE_NAME}"
+    if [[ ${BUILDING_STATIC_FRAMEWORK:=0} == 0 ]]; then
+        cp -RP "${FRAMEWORK_ARM_BUILD_PATH}" "${FRAMEWORK_DESTINATION_PATH}"
+        xcrun lipo -create "${FRAMEWORK_DESTINATION_PATH}/${frameworkName}" "${FRAMEWORK_SIM_BUILD_PATH}/${frameworkName}" -output "${FRAMEWORK_DESTINATION_PATH}/${frameworkName}"
+    else
+        [[ ! -d "${FRAMEWORK_PRODUCTS_PATH}" ]] && mkdir -p "${FRAMEWORK_PRODUCTS_PATH}"
+        if [[ ! -d "${FRAMEWORK_PRODUCTS_PATH}/Headers" ]]; then
+            cp -RP "${FRAMEWORK_ARM_BUILD_PATH}/Headers" "${FRAMEWORK_PRODUCTS_PATH}/Headers"
+            cp "${FRAMEWORK_ARM_BUILD_PATH}/PubNub-iOS-Info.plist" "${FRAMEWORK_PRODUCTS_PATH}/Info.plist"
+        fi
+        cp "${FRAMEWORK_ARM_BUILD_PATH}/${frameworkName}.a" "${FRAMEWORK_PRODUCTS_PATH}/${frameworkName}.a"
+
+        xcrun lipo -create "${FRAMEWORK_PRODUCTS_PATH}/${frameworkName}.a" "${FRAMEWORK_SIM_BUILD_PATH}/${frameworkName}.a" -output "${FRAMEWORK_PRODUCTS_PATH}/${frameworkName}"
+        rm "${FRAMEWORK_PRODUCTS_PATH}/${frameworkName}.a"
+    fi
+
+    if [[ ${BUILDING_STATIC_FRAMEWORK:=0} == 0 ]]; then
+        cp -RP "${FRAMEWORKS_PATH}/${FRAMEWORK_BUNDLE_NAME}" "${FRAMEWORK_PRODUCTS_PATH}"
+    fi
 done
